Add buffer based tests for draw_rect
At this point the tests of fill functionality will and ought to fail.

The test cases are bundled in to an array which is passed to a single
test function.

It may be preferrable to make use of a testing crate, e.g rstest, and to
split out each test in to a separate test, but
see
https://github.com/ratatui-org/ratatui/blob/359204c9298cc26ea21807d886d596de0329bacc/tests/widgets_chart.rs#L66
for an alternative approach from the exemplar of buffer testing.f